refactor now that the codebase is bigger, can now download files from minio
